Add component to display a progress bar
