+++ less/components/loading-bar.less
@@ -1,0 +1,30 @@
+.LoadingBar {
+    pointer-events: none;
+    transition: 400ms linear all;
+
+    .bar {
+        background: @gb-brand-primary;
+        height: 2px;
+
+        position: fixed;
+        top: 0;
+        left: 0;
+        z-index: 10000;
+        display: none;
+        width: 100%;
+        border-radius: 0 1px 1px 0;
+        transition: width 350ms;
+    }
+
+    .LoadingBar-shadow {
+        content: '';
+        position: absolute;
+        top: 0;
+        right: 0;
+        width: 70px;
+        height: 2px;
+        border-radius: 50%;
+        opacity: .45;
+        box-shadow: @gb-brand-primary 1px 0 6px 1px;
+    }
+}
