+++ less/main.less
@@ -38,6 +38,7 @@
 @import "./components/pagination.less";
 @import "./components/panels.less";
 @import "./components/popover.less";
+@import "./components/loading-bar.less";
 @import "./components/progress-bar.less";
 @import "./components/search-bar.less";
 @import "./components/select.less";
