+++ less/variables.less
@@ -382,6 +382,11 @@
 @gb-tree-selected-background:    #dcdfe1;
 
 ///
+/// Loading bar
+///
+@gb-loading-bar-background:      #f5f5f5;
+
+///
 /// Cards
 ///
 @gb-card-border:                  @gb-palette-gray-outline;
